+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1.25 (in development)
 
+- Added `AspectRatio` helper class
 - Improved YouTube embed URL detection in `EmbedUrlTrait`
 
 ## 2.1.24 (Oct 1, 2024)